# Faster indexing math in a few kernels