… docs/src/usage/function_transforms.rst
… mlx/backend/metal/copy.cpp
… mlx/backend/metal/indexing.cpp
… mlx/backend/metal/kernels/binary.h
… mlx/backend/metal/kernels/binary.metal
… mlx/backend/metal/kernels/binary_two.h
… mlx/backend/metal/kernels/binary_two.metal
… mlx/backend/metal/kernels/copy.h
… mlx/backend/metal/kernels/copy.metal
… mlx/backend/metal/kernels/gather.h
… mlx/backend/metal/kernels/scatter.h
… mlx/backend/metal/kernels/ternary.h
… mlx/backend/metal/kernels/ternary.metal
… mlx/backend/metal/kernels/unary.h
… mlx/backend/metal/kernels/unary.metal
… mlx/backend/metal/kernels/utils.h
… mlx/backend/metal/ternary.cpp
… mlx/backend/metal/unary.cpp
… mlx/backend/metal/utils.h