- Use `uint` instead of `size_t` in unary, binary, ternary, copy, and jit compile kernels when possible.

- Generally changing code that I touch from using `ostringstream` to `string` concatenation / formatting which prior benchmarking shows can be a lot faster. I also think it can be more readable.

### Microbenchmarks on M1 Max:

Bench | Pre (ms) | Post (ms)
----- | ---- | ----
Unary general 5D | 4.721 | 2.950 
Binary general 3D | 6.164 |  4.761
Binary general 5D | 6.147 |  5.947
Ternary general 3D | 6.209 |  5.236 
Compiled batch norm | 4.784 | 3.171 
Gather 3D |3.913 | 3.105
Scatter 5D | 43.536 | 10.827

### Microbenchmarks on M3 Max:

Bench | Pre (ms) | Post (ms)
----- | ---- | ----
Unary general 5D | 3.412 | 3.093
Binary general 3D | 3.063 | 3.032
Binary general 5D | 5.989 |  5.828
Ternary general 3D | 3.351 |  3.339
Compiled batch norm | 3.66 | 2.62

### CIFAR ResNet Training benchmark on M3 Max

Pre: 28.2k ims/sec
Post 34.2k ims/sec

### No JIT binary sizes:

Pre: uncompressed 69M, compressed 19M
Post: uncompressed 77M, compressed 19M