@awni 2024-11-19
> It is interesting that work per thread of 2 is faster than 1 but presumably slower than 4 for uint indexing.

It is interesting indeed. I tested all the ones I changed for at least one case mostly on an M1 Max.. but probably there is some variability based on the actual op and machine. 

In general more threads seems to be better unless we hit against indexing being expensive then we have to amortize the cost.